requirements.txt
aiofiles==0.7.0
aniso8601==9.0.1
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
backports.zoneinfo==0.2.1
bcrypt==3.2.0
black==21.10b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==36.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.58.1
graphene==3.0
graphql-core==3.1.6
graphql-relay==3.1.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
orjson==3.6.4
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.4
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.0
starlette==0.13.4
tomli==1.2.2
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
